chore: rename folder to `parsers`
#121
