+++ src/parsers/endOfInput.test.ts
