+++ src/parsers/endOfLine.test.ts
