+++ src/parsers/endOfLine.ts
