+++ src/parsers/lazy.ts
