+++ src/parsers/string.test.ts
@@ -1,9 +1,9 @@
-import { regex } from "../parser/index";
+import { string } from "./string";
 import { ParserState } from "../ParserState";
 
-describe("regex", () => {
+describe("string", () => {
   test("Parse a matchable input", () => {
-    const parser = regex(/hello/i);
+    const parser = string("Hello");
 
     const result = parser.run("Hello");
 
@@ -16,14 +16,14 @@ describe("regex", () => {
   });
 
   test("Parse a not matchable input", () => {
-    const parser = regex(/hello/i);
+    const parser = string("Hello");
 
     const result = parser.run("World");
 
     expect(result).toStrictEqual<ParserState>({
       input: "World",
       isError: true,
-      errorMessage: "Failed to match regex /hello/i at offset 0",
+      errorMessage: "Failed to match string Hello at offset 0",
       offset: 0,
       result: null,
     });
