+++ src/parsers/string.ts
