+++ src/parsers/word.ts
